Update Review.java
the "totalSentiment" method takes a file, converts its text into a string, then scans through the length of the string. While scanning, it places each letter into a placeholder String and looks for spaces (since a space means the start of a new word). Then, it removes punctuation to ensure true congruency between the placeholder String and the corresponding adjective String. Then, the total sentiment of that word is added to a totalSentiment double, the placeholder String is reset, and that process repeats until the whole String is exhausted. Finally, the totalSentiment is returned as a double.