fakeReview Method
The fakeReview method works by taking a file's name as a parameter and unpacking that file into a String. Then, a for loop scans through the String looking for an asterisk which indicates an adjective, while also adding each character into a placeholder String. The adjective once detected by an asterisk is not placed into the placeholder String. Then, once it reaches a space, a random adjective using the randomAdjective method replaces the passed-through adjective. This process repeats until the entire String is exhausted and then the placeholder String which now has the same String but with randomized adjectives is returned.