fakeReviewStronger Method
This method is similar to the fakeReview method, however it has a small caveat. Once it detects the adjective, it starts to build a seperate String that houses the adjective. Once the adjective is complete, indicated by a space being reached, a while loop is initiated. This while loop stays true until a randomly generated adjective is "stronger" than the adjective being tested. So, if an adjective has a negative sentiment value, it will look for adjectives that have a negative sentiment value and that are more negative than the tested adjective. The same applies to a positive adjective but in a "positively stronger" way. Finally, once the adjective is found, the while loop breaks, the adjective is added to the placeholder String, and the asteriskDetected boolean and adjective String are reset.