Added a removePunctuation method
Joe brought an issue to my attention in person, the adjective may have a comma next to it which leads to the sentimentVal method not returning a proper value. Therefore, I added a removePunctuation method which takes the adjective, which may or may not have any punctuation, and removes any punctuation in the beginning and end of it. That allows it to filter through the rest of the if statement without being obstructed.