Finishing Touches
I noticed a strange bug with the fakeReview and fakeReviewStronger method during my final tests. It would sometimes have an adjective be returned as "" for a reason that I had no idea about. So, I added a small change for the fakeReview and fakeReviewStronger method that takes the randomly generated adjective, and does not allow it to be returned or go any further if it equals "". This should make these methods without any significant bugs now.