Fixed Punctuation
I added a String that holds typical punctuation characters. This String is used in an indexOf method to see if the character prior to the space after an adjective is a piece of punctuation. If the indexOf method returns anything other than -1, that character exists within the punctuation String and therefore is a piece of punctuation to be added to the placeholder String. Otherwise, it just adds a space and acts like nothing ever happened.